requirements.txt
asgiref==3.5.2
coverage==6.4
dataclasses-json==0.5.7
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-schema-adapter@ https://github.com/flightsafe/drf-schema-adapter/archive/refs/tags/2.1.1.zip
future==0.18.2
gunicorn==20.1.0
Inflector==3.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==21.3
Pillow==9.1.1
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
serializers==0.2.4
sqlparse==0.4.2
typing-inspect==0.7.1
typing_extensions==4.2.0
uritemplate==4.1.1
